==> scripts/benchmark/03_benchmark_weights.ipynb <==
# Evaluation of binding weights in CollecTRI

## Libraries

import pandas as pd
import numpy as np

import decoupler as dc
import matplotlib.pyplot as plt

## Benchmark data

### Load KnockTF data
The benchmark data can be downloaded from Zenodo as shown in scripts/benchmark.ipynb

mat = pd.read_csv('../../data/knockTF_expr.csv', index_col=0)
obs = pd.read_csv('../../data/knockTF_meta.csv', index_col=0)

### Filter knockTF data
Filtering is based on log fold-change of perturbed transcription factors

msk = obs['logFC'] < -1
mat = mat[msk]
obs = obs[msk]
mat.shape, obs.shape

## Load weighted GRNs

matRid1raw = pd.read_csv('../../output/weighted_networks/matrixRider_1000bp_raw.csv')
matRid1gene = pd.read_csv('../../output/weighted_networks/matrixRider_1000bp_geneNorm.csv')
matRid1tf = pd.read_csv('../../output/weighted_networks/matrixRider_1000bp_tfNorm.csv')

matRid10raw = pd.read_csv('../../output/weighted_networks/matrixRider_10000bp_raw.csv')
matRid10gene = pd.read_csv('../../output/weighted_networks/matrixRider_10000bp_geneNorm.csv')
matRid10tf = pd.read_csv('../../output/weighted_networks/matrixRider_10000bp_tfNorm.csv')

FIMO1raw = pd.read_csv('../../output/weighted_networks/FIMO_1000bp_raw.csv')
FIMO1gene = pd.read_csv('../../output/weighted_networks/FIMO_1000bp_geneNorm.csv')
FIMO1tf = pd.read_csv('../../output/weighted_networks/FIMO_1000bp_tfNorm.csv')

FIMO10raw = pd.read_csv('../../output/weighted_networks/FIMO_10000bp_raw.csv')
FIMO10gene = pd.read_csv('../../output/weighted_networks/FIMO_10000bp_geneNorm.csv')
FIMO10tf = pd.read_csv('../../output/weighted_networks/FIMO_10000bp_tfNorm.csv')

collecTRI = pd.read_csv('../../output/weighted_networks/matrixRider_1000bp_raw.csv')
collecTRI['weight'] = np.where(collecTRI['weight'] >= 0, 1, -1)


## Run benchmark weighted CollecTRI

# Build dictionary of networks to test
nets = {
    'collecTRI': collecTRI,
    'matRid1raw': matRid1raw,
    'matRid1gene': matRid1gene,
    'matRid1tf': matRid1tf,
    'matRid10raw': matRid10raw,
    'matRid10gene': matRid10gene,
    'matRid10tf': matRid10tf,
    'FIMO1raw': FIMO1raw,
    'FIMO1gene': FIMO1gene,
    'FIMO1tf': FIMO1tf,
    'FIMO10raw': FIMO10raw,
    'FIMO10gene': FIMO10gene,
    'FIMO10tf': FIMO10tf
}

# Example extra arguments
decouple_kws = {
    'collecTRI': {'methods': 'ulm'},
    'matRid1raw': {'methods': 'ulm'},
    'matRid1gene': {'methods': 'ulm'},
    'matRid1tf': {'methods': 'ulm'},
    'matRid10raw': {'methods': 'ulm'},
    'matRid10gene': {'methods': 'ulm'},
    'matRid10tf': {'methods': 'ulm'},
    'FIMO1raw': {'methods': 'ulm'},
    'FIMO1gene': {'methods': 'ulm'},
    'FIMO1tf': {'methods': 'ulm'},
    'FIMO10raw': {'methods': 'ulm'},
    'FIMO10gene': {'methods': 'ulm'},
    'FIMO10tf': {'methods': 'ulm'}
}

# Run benchmark pipeline
df_weights = dc.benchmark(mat, obs, nets, perturb='TF', sign=-1, decouple_kws=decouple_kws)

### Save data

pd.DataFrame.to_csv(df_weights, '../../output/benchmark/benchmark_weights_res.csv')

## Load filtered GRNs based on weights 

matRid1raw = pd.read_csv('../../output/weighted_networks/matrixRider_1000bp_raw.csv')
matRid1raw['weight'] = np.where(matRid1raw['weight'] >= 0, 1, -1)
matRid1_10 = pd.read_csv('../../output/weighted_networks/matrixRider_1000bp_filtered10.csv')
matRid1_20 = pd.read_csv('../../output/weighted_networks/matrixRider_1000bp_filtered20.csv')
matRid1_30 = pd.read_csv('../../output/weighted_networks/matrixRider_1000bp_filtered30.csv')

matRid10raw = pd.read_csv('../../output/weighted_networks/matrixRider_10000bp_raw.csv')
matRid10raw['weight'] = np.where(matRid10raw['weight'] >= 0, 1, -1)
matRid10_10 = pd.read_csv('../../output/weighted_networks/matrixRider_10000bp_filtered10.csv')
matRid10_20 = pd.read_csv('../../output/weighted_networks/matrixRider_10000bp_filtered20.csv')
matRid10_30 = pd.read_csv('../../output/weighted_networks/matrixRider_10000bp_filtered30.csv')

FIMO1raw = pd.read_csv('../../output/weighted_networks/FIMO_1000bp_raw.csv')
FIMO1raw['weight'] = np.where(FIMO1raw['weight'] >= 0, 1, -1)
FIMO1_10 = pd.read_csv('../../output/weighted_networks/FIMO_1000bp_filtered10.csv')
FIMO1_20 = pd.read_csv('../../output/weighted_networks/FIMO_1000bp_filtered20.csv')
FIMO1_30 = pd.read_csv('../../output/weighted_networks/FIMO_1000bp_filtered30.csv')

FIMO10raw = pd.read_csv('../../output/weighted_networks/FIMO_10000bp_raw.csv')
FIMO10raw['weight'] = np.where(FIMO10raw['weight'] >= 0, 1, -1)
FIMO10_10 = pd.read_csv('../../output/weighted_networks/FIMO_10000bp_filtered10.csv')
FIMO10_20 = pd.read_csv('../../output/weighted_networks/FIMO_10000bp_filtered20.csv')
FIMO10_30 = pd.read_csv('../../output/weighted_networks/FIMO_10000bp_filtered30.csv')

collecTRI = pd.read_csv('../../output/weighted_networks/matrixRider_1000bp_raw.csv')
collecTRI['weight'] = np.where(matRid1raw['weight'] >= 0, 1, -1)
collecTRI_10 = pd.read_csv('../../output/weighted_networks/random10.csv')
collecTRI_20 = pd.read_csv('../../output/weighted_networks/random20.csv')
collecTRI_30 = pd.read_csv('../../output/weighted_networks/random30.csv')


## Run benchmark of filtered CollecTRI

# Build dictionary of networks to test
nets = {
    'matRid1raw': matRid1raw,
    'matRid1_10': matRid1_10,
    'matRid1_20': matRid1_20,
    'matRid1_30': matRid1_30,
    'matRid10raw': matRid10raw,
    'matRid10_10': matRid10_10,
    'matRid10_20': matRid10_20,
    'matRid10_30': matRid10_30,
    'FIMO1raw': FIMO1raw,
    'FIMO1_10': FIMO1_10,
    'FIMO1_20': FIMO1_20,
    'FIMO1_30': FIMO1_30,
    'FIMO10raw': FIMO10raw,
    'FIMO10_10': FIMO10_10,
    'FIMO10_20': FIMO10_20,
    'FIMO10_30': FIMO10_30,
    'collecTRI': collecTRI,
    'collecTRI_10': collecTRI_10,
    'collecTRI_20': collecTRI_20,
    'collecTRI_30': collecTRI_30
}

# Example extra arguments
decouple_kws = {
    'matRid1raw': {'methods': 'ulm'},
    'matRid1_10': {'methods': 'ulm'},
    'matRid1_20': {'methods': 'ulm'},
    'matRid1_30': {'methods': 'ulm'},
    'matRid10raw': {'methods': 'ulm'},
    'matRid10_10': {'methods': 'ulm'},
    'matRid10_20': {'methods': 'ulm'},
    'matRid10_30': {'methods': 'ulm'},
    'FIMO1raw': {'methods': 'ulm'},
    'FIMO1_10': {'methods': 'ulm'},
    'FIMO1_20': {'methods': 'ulm'},
    'FIMO1_30': {'methods': 'ulm'},
    'FIMO10raw': {'methods': 'ulm'},
    'FIMO10_10': {'methods': 'ulm'},
    'FIMO10_20': {'methods': 'ulm'},
    'FIMO10_30': {'methods': 'ulm'},
    'collecTRI': {'methods': 'ulm'},
    'collecTRI_10': {'methods': 'ulm'},
    'collecTRI_20': {'methods': 'ulm'},
    'collecTRI_30': {'methods': 'ulm'}
}

# Run benchmark pipeline
df_filtered = dc.benchmark(mat, obs, nets, perturb='TF', sign=-1, verbose=True, decouple_kws=decouple_kws)

### Save data

pd.DataFrame.to_csv(df_filtered, '../../output/benchmark/benchmark_weights_filtered_res.csv')
